requirements-dev.txt
# Development dependencies for Access Log Analyzer

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.0

# Type stubs
pandas-stubs>=2.1.0
types-PyYAML>=6.0.0
